Refactor: 폴더 구조 변경 - inputs, buttons 폴더 추가
