+++ app/page.tsx
@@ -1,5 +1,5 @@
+import SwitchThemeButton from '@/components/buttons/switchTheme/SwitchThemeButton';
 import SignupForm from '@/components/signupForm/SignupForm';
-import SwitchThemeButton from '@/components/switchTheme/SwitchThemeButton';
 
 export default function Home() {
   return (
