+++ components/buttons/switchTheme/SwitchThemeButton.tsx
@@ -5,13 +5,13 @@ import React from 'react';
 import { Moon, Sun } from 'lucide-react';
 import { useTheme } from 'next-themes';
 
-import { Button } from '../ui/button';
+import { Button } from '../../ui/button';
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from '../ui/dropdown';
+} from '../../ui/dropdown';
 
 const SwitchThemeButton = () => {
   const { setTheme } = useTheme();
